+++ crates/dwui/src/components/input/mod.rs
@@ -1,4 +1,6 @@
+pub mod calendar_date;
 pub mod checkbox;
+pub mod date_picker;
 pub mod select;
 pub mod slider;
 pub mod switch;
@@ -6,7 +8,9 @@ pub mod text_input;
 pub mod validation;
 
 pub mod prelude {
+    pub use super::calendar_date::*;
     pub use super::checkbox::*;
+    pub use super::date_picker::*;
     pub use super::select::*;
     pub use super::slider::*;
     pub use super::switch::*;